backend-resume-anaalyser/requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
